refactor: exportable-manifest
