+++ packages/exportable-manifest/test/index.test.ts
@@ -1,6 +1,5 @@
 /// <reference path="../../../typings/index.d.ts"/>
-import exportableManifest, { overridePublishConfig } from '@pnpm/exportable-manifest'
-import { PackageManifest, PublishConfig } from '@pnpm/types'
+import exportableManifest from '@pnpm/exportable-manifest'
 
 test('the pnpm options are removed', async () => {
   expect(await exportableManifest(process.cwd(), {
@@ -55,34 +54,5 @@ test('readme added to published manifest', async () => {
     name: 'foo',
     version: '1.0.0',
     readme: 'readme content',
-  })
-})
-
-test('publish config to be overridden', async () => {
-  const publishConfig: PublishConfig = {
-    main: 'overridden',
-    types: 'overridden',
-    typesVersions: {
-      '*': {
-        '*': ['overridden'],
-      },
-    },
-  }
-  const publishManifest: PackageManifest = {
-    name: 'foo',
-    version: '1.0.0',
-    main: 'origin',
-    types: 'origin',
-    typesVersions: {
-      '*': {
-        '*': ['origin'],
-      },
-    },
-    publishConfig,
-  }
-  overridePublishConfig(publishManifest)
-
-  Object.keys(publishConfig).forEach((publishConfigKey) => {
-    expect(publishManifest[publishConfigKey]).toEqual(publishConfig[publishConfigKey])
   })
 })
